maint(tests): add proper integration tests
